gui update + death screen
I added pictures to all of the screens, and changed the fonts and moved some stuff around. I also added a death screen, it doesn't have the death image yet but i'll add that later. I put everything into a new python file called gabi's code just in case so it wouldn't mess up what you'd done already. I also made an images folder.

For the pictures to work you have to download pillow into the typeIt repository. Not sure if it'll automatically be on urs cuz i added it to mine