level 2 issues
i tried to fix level 2 since it seemed like it wasn't working. i changed it in the gabi's code file. seems to run now cuz it wasn't before, but theres a problem where if u wanna capitalize something and u press shift, it counts that as a wrong button pressed and you die. so idk how to fix that